requirements.txt
aiogram~=3.18.0
asyncpg
sqlalchemy[asyncio]~=2.0.38
pydantic~=2.10.6
pydantic-settings~=2.8.0
alembic[asyncio]

python-dotenv~=1.0.1
requests
python-crontab